\name{tourData09}
\Rdversion{1.1}
\docType{data}
\alias{tourData09}
\title{Tour de France 2009}
\description{Each row represents the performance of a rider in the
  2009 Tour de France; the name and team of the rider are used as the
  row names. The four columns are the \code{Cost} (to include on a team
  in the Versus fanatasy challenge), \code{Scores} (based on daily
  finishing position), \code{JerseyBonus} (for any days spent in one of
  the three main leader jerseys), and \code{Total} (the sum of Scores
  and JerseyBonus).
}
\usage{data(tourData09)}
\format{A data frame with 102 rows and 4 columns.}
\source{\url{http://www.versus.com/tdfgames}}
\references{
  Coombes KR. Genetic Algorthims, Feature Selection, and the Tour de
  France. Submitted.
}
\keyword{datasets}
